requirements.txt
h5py==2.7.1
imageio==2.2.0
ipython==6.2.1
ipython-genutils==0.2.0
Markdown==2.6.11
matplotlib==2.1.2
numpy==1.14.0
pandas==0.22.0
Pillow==4.3.0
pynrrd==0.2.4
scikit-learn==0.19.1
scipy==1.0.0
sklearn==0.0
tensorflow==1.4.1
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc3
